refactor: 清理和优化拜访好友功能
- 移除不再使用的自定义识别器和动作，简化代码结构。
- 在 `select_friend.go` 中新增 `normalizeFriendName` 函数以处理好友名称的清洗逻辑。
- 更新 `.vscode/settings.json` 以设置 JSONC 格式的默认格式化工具。
- 调整多个 JSON 配置文件，移除冗余的动作和识别器，确保功能的准确性和一致性。

这些改动旨在提升代码可维护性和功能的清晰度。
