+++ assets/resource/pipeline/VisitFriends/MenuScan.json
@@ -99,83 +99,6 @@
             "Node.Action.Succeeded": "向下滑动好友列表"
         }
     },
-    "VisitFriendsMenuScanDetail": {
-        "desc": "识别进入好友船的按钮",
-        "pre_delay": 0,
-        "post_delay": 0,
-        "rate_limit": 0,
-        "next": [
-            "VisitFriendsMenuScanDetailClueExchange"
-        ]
-    },
-    "VisitFriendsMenuScanDetailClueExchange": {
-        "desc": "识别是否可以线索交换",
-        "recognition": "Custom",
-        "custom_recognition": "VisitFriendsMenuScanDetailClueExchangeRecognition",
-        "custom_recognition_param": {
-            "friend_name": "aaa#123",
-            "enter_button_box": [
-                0,
-                0,
-                100,
-                100
-            ]
-        },
-        "pre_delay": 0,
-        "post_delay": 0,
-        "rate_limit": 0,
-        "next": [
-            "VisitFriendsMenuScanDetailOpen",
-            "VisitFriendsMenuScanDetailFinish"
-        ]
-    },
-    "VisitFriendsMenuScanDetailOpen": {
-        "desc": "识别进入好友船的按钮",
-        "pre_delay": 0,
-        "action": "Click",
-        "target": [
-            -100,
-            0,
-            100,
-            100
-        ],
-        "target_offset": [
-            5,
-            5,
-            -10,
-            -10
-        ],
-        "post_delay": 0,
-        "next": [
-            "VisitFriendsMenuScanDetailOpenSuccess"
-        ]
-    },
-    "VisitFriendsMenuScanDetailOpenSuccess": {
-        "desc": "成功打开好友详情弹窗",
-        "recognition": "And",
-        "all_of": [
-            "GrayCancelButton"
-        ],
-        "pre_wait_freezes": 100,
-        "pre_delay": 0,
-        "post_delay": 0,
-        "next": [
-            "VisitFriendsMenuScanDetailSaveAssist"
-        ]
-    },
-    "VisitFriendsMenuScanDetailSaveAssist": {
-        "desc": "保存好友船助力详情（旧路径，待 CheckAssist 完全替换后删除）",
-        "recognition": "Custom",
-        "custom_recognition": "VisitFriendsMenuScanDetailAssistRecognition",
-        "custom_recognition_param": {
-            "friend_name": "aaa#1234"
-        },
-        "pre_delay": 0,
-        "post_delay": 0,
-        "next": [
-            "VisitFriendsMenuScanDetailClose"
-        ]
-    },
     "VisitFriendsCheckAssist": {
         "desc": "打开好友详情后检查可访问资源（识别②）：线索交换固定开启；助力子节点未勾选时为 OCR a^，勾选后覆盖为 TemplateMatch",
         "recognition": "Or",
@@ -197,7 +120,7 @@
         ]
     },
     "VisitFriendsCheckAssistClueExchange": {
-        "desc": "检查是否可线索交换",
+        "desc": "检查是否可线索交换（固定开启，不可关闭）",
         "recognition": "And",
         "all_of": [
             "VisitFriendsRecognitionItemClueExchange"
@@ -233,11 +156,6 @@
         "pre_delay": 0,
         "action": "Click",
         "post_wait_freezes": 400,
-        "post_delay": 0
-    },
-    "VisitFriendsMenuScanDetailFinish": {
-        "desc": "完成好友详情识别",
-        "pre_delay": 0,
         "post_delay": 0
     }
 }
