+++ tools/schema/custom.action.schema.json
@@ -79,13 +79,6 @@
                 "SceneManagerMenuListClickItemAction",
                 "SubTask",
                 "TrialOfSwordmancy.Decide",
-                "VisitFriendsExecuteVisitAction",
-                "VisitFriendsMainAction",
-                "VisitFriendsMenuClueAssistAction",
-                "VisitFriendsMenuClueAssistFullAction",
-                "VisitFriendsMenuClueExchangeAction",
-                "VisitFriendsMenuClueExchangeFullAction",
-                "VisitFriendsMenuScanTargetFriendOpenAction",
                 "WebEvent202605Action",
                 "autoEcoFarmInterruptibleSleep",
                 "autoEcoFarmOverrideTargetTemplate",
