+++ tools/schema/custom.recognition.schema.json
@@ -27,12 +27,7 @@
                 "ScheduleRecognition",
                 "SellProductNormalizedItemMatch",
                 "TrialOfSwordmancy.Recognize",
-                "VisitFriendsMenuScanDetailAssistRecognition",
-                "VisitFriendsMenuScanDetailClueExchangeRecognition",
-                "VisitFriendsMenuScanScrollFinishRecognition",
-                "VisitFriendsMenuScanScrollFullRecognition",
                 "VisitFriendsSelectFriendRecognition",
-                "VisitFriendsMenuScanTargetFriendOpenRecognition",
                 "WeaponInventoryScanRecognition",
                 "SeizeDeliveryJobsScanTargetRecognition",
                 "autoEcoFarmCalculateSwipeTarget",
